requirements.txt
Flask
Flask-Mail
SQLAlchemy
python-dotenv
itsdangerous
werkzeug
